# speed up listDatasetFileDetails API