--- a/src/python/WMCore/Services/DBS/DBS3Reader.py
+++ b/src/python/WMCore/Services/DBS/DBS3Reader.py
@@ -23,6 +23,8 @@ from retry import retry
 from Utils.IteratorTools import grouper
 from Utils.PythonVersion import PY2
 from WMCore.Services.DBS.DBSErrors import DBSReaderError, formatEx3
+from WMCore.Services.DBS.DBSUtils import dbsListFileParents, dbsListFileLumis, \
+    dbsBlockOrigin, dbsParentFilesGivenParentDataset
 
 
 ### Needed for the pycurl comment, leave it out for now
@@ -72,17 +74,28 @@ class DBS3Reader(object):
     General API for reading data from DBS
     """
 
-    def __init__(self, url, logger=None, **contact):
+    def __init__(self, url, logger=None, parallel=None, **contact):
+        """
+        DBS3Reader constructor
+
+        :param url: url of DBS server
+        :param logger: logger to be used by this class
+        :param parallel: optional parameter to specify parallel execution of some APIs
+        You may pass any true value, e.g. True or 1. The parallel APIs are:
+        listDatasetFileDetails, listFileBlockLocation, getParentFilesGivenParentDataset
+        :param contact: optional parameters to pass to DbsApi class
+        """
 
         # instantiate dbs api object
         try:
             self.dbsURL = url.replace("cmsweb.cern.ch", "cmsweb-prod.cern.ch")
             self.dbs = DbsApi(self.dbsURL, **contact)
             self.logger = logger or logging.getLogger(self.__class__.__name__)
+            self.parallel = parallel
         except Exception as ex:
             msg = "Error in DBSReader with DbsApi\n"
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
     def _getLumiList(self, blockName=None, lfns=None, validFileOnly=1):
         """
@@ -103,7 +116,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.listFileLumiArray(%s)\n" % lfns
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         lumiDict = {}
         for lumisItem in lumiLists:
@@ -128,7 +141,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBS server is not up: %s" % self.dbsURL
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
     def listPrimaryDatasets(self, match='*'):
         """
@@ -143,7 +156,7 @@ class DBS3Reader(object):
         except Exception as ex:
             msg = "Error in DBSReader.listPrimaryDataset(%s)\n" % match
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         result = [x['primary_ds_name'] for x in result]
         return result
@@ -160,7 +173,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader.listProcessedDatasets(%s)\n" % primary
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         for dataset in datasets:
             dataset = remapDBS3Keys(dataset, processed_ds_name='Name')
@@ -195,7 +208,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader.listRuns(%s, %s)\n" % (dataset, block)
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
         for x in results:
             runs.extend(x['run_num'])
         return runs
@@ -228,7 +241,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader.listRuns(%s, %s)\n" % (dataset, block)
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         # send runDict format as result, this format is for sync with dbs2 call
         # which has {run_number: num_lumis} but dbs3 call doesn't return num Lumis
@@ -254,7 +267,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader.listProcessedDatasets(%s)\n" % primary
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         result = [x['dataset'].split('/')[2] for x in result]
         return result
@@ -307,6 +320,25 @@ class DBS3Reader(object):
             files[f['logical_file_name']]['Lumis'] = {}
             files[f['logical_file_name']]['Parents'] = []
 
+        # parallel execution for listFileParents and listFileLumis APIs
+        if self.parallel:
+            if getParents:
+                block_parents = dbsListFileParents(self.dbsURL, blocks)
+                for blockName, parents in block_parents.items():
+                    for p in parents:
+                        if p['logical_file_name'] in files:  # invalid files are not there if validFileOnly=1
+                            files[p['logical_file_name']]['Parents'].extend(p['parent_logical_file_name'])
+            if getLumis:
+                block_file_lumis = dbsListFileLumis(self.dbsURL, blocks)
+                for blockName, file_lumis in block_file_lumis.items():
+                    for f in file_lumis:
+                        if f['logical_file_name'] in files:  # invalid files are not there if validFileOnly=1
+                            if f['run_num'] in files[f['logical_file_name']]['Lumis']:
+                                files[f['logical_file_name']]['Lumis'][f['run_num']].extend(f['lumi_section_num'])
+                            else:
+                                files[f['logical_file_name']]['Lumis'][f['run_num']] = f['lumi_section_num']
+            return files
+
         # Iterate over the blocks and get parents and lumis
         for blockName in blocks:
             # get the parents
@@ -344,7 +376,7 @@ class DBS3Reader(object):
         except Exception as exc:
             msg = "Error in DBSReader.crossCheck({}) with {} lfns.".format(datasetPath, len(lfns))
             msg += "\nDetails: {}\n".format(formatEx3(exc))
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
         setOfAllLfns = set(allLfns)
         setOfKnownLfns = set(lfns)
         return list(setOfAllLfns.intersection(setOfKnownLfns))
@@ -363,7 +395,7 @@ class DBS3Reader(object):
         except Exception as exc:
             msg = "Error in DBSReader.crossCheckMissing({}) with {} lfns.".format(datasetPath, len(lfns))
             msg += "\nDetails: {}\n".format(formatEx3(exc))
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
         setOfAllLfns = set(allLfns)
         setOfKnownLfns = set(lfns)
         knownFiles = setOfAllLfns.intersection(setOfKnownLfns)
@@ -384,7 +416,7 @@ class DBS3Reader(object):
         except Exception as ex:
             msg = "Error in DBSReader.getDBSSummaryInfo(%s, %s)\n" % (dataset, block)
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         if not summary:  # missing data or all files invalid
             return {}
@@ -411,7 +443,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader.listFileBlocks(%s)\n" % dataset
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         result = [x['block_name'] for x in blocks]
 
@@ -435,7 +467,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.blockExists(%s)\n" % fileBlockName
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         if len(blocks) == 0:
             return False
@@ -454,7 +486,7 @@ class DBS3Reader(object):
         """
         if not self.blockExists(fileBlockName):
             msg = "DBSReader.listFilesInBlock(%s): No matching data"
-            raise DBSReaderError(msg % fileBlockName)
+            raise DBSReaderError(msg % fileBlockName) from None
 
         try:
             files = self.dbs.listFileArray(block_name=fileBlockName, validFileOnly=validFileOnly, detail=True)
@@ -462,7 +494,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.listFilesInBlock(%s)\n" % fileBlockName
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         if lumis:
             lumiDict = self._getLumiList(blockName=fileBlockName, validFileOnly=validFileOnly)
@@ -487,7 +519,7 @@ class DBS3Reader(object):
         """
         if not self.blockExists(fileBlockName):
             msg = "DBSReader.listFilesInBlockWithParents(%s): No matching data"
-            raise DBSReaderError(msg % fileBlockName)
+            raise DBSReaderError(msg % fileBlockName) from None
 
         try:
             # TODO: shoud we get only valid block for this?
@@ -499,7 +531,7 @@ class DBS3Reader(object):
             msg += "DBSReader.listFilesInBlockWithParents(%s)\n" % (
                 fileBlockName,)
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         childByParents = defaultdict(list)
         for f in files:
@@ -513,7 +545,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.listFilesInBlockWithParents(%s)\n There is no parents files" % (
                 fileBlockName)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         parentFilesDetail = []
         # TODO: slicing parentLFNs util DBS api is handling that.
@@ -549,7 +581,7 @@ class DBS3Reader(object):
         """
         if not self.blockExists(fileBlockName):
             msg = "DBSReader.lfnsInBlock(%s): No matching data"
-            raise DBSReaderError(msg % fileBlockName)
+            raise DBSReaderError(msg % fileBlockName) from None
 
         try:
             lfns = self.dbs.listFileArray(block_name=fileBlockName, validFileOnly=1, detail=False)
@@ -558,7 +590,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.listFilesInBlock(%s)\n" % fileBlockName
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
     def listFileBlockLocation(self, fileBlockNames):
         """
@@ -581,15 +613,22 @@ class DBS3Reader(object):
 
         blocksInfo = {}
         try:
-            for block in fileBlockNames:
-                blocksInfo.setdefault(block, [])
-                # there should be only one element with a single origin site string ...
-                for blockInfo in self.dbs.listBlockOrigin(block_name=block):
-                    blocksInfo[block].append(blockInfo['origin_site_name'])
+            if self.parallel:
+                data = dbsBlockOrigin(self.dbsURL, fileBlockNames)
+                for block, items in data.items():
+                    blocksInfo.setdefault(block, [])
+                    for blockInfo in items:
+                        blocksInfo[block].append(blockInfo['origin_site_name'])
+            else:
+                for block in fileBlockNames:
+                    blocksInfo.setdefault(block, [])
+                    # there should be only one element with a single origin site string ...
+                    for blockInfo in self.dbs.listBlockOrigin(block_name=block):
+                        blocksInfo[block].append(blockInfo['origin_site_name'])
         except dbsClientException as ex:
             msg = "Error in DBS3Reader: self.dbs.listBlockOrigin(block_name=%s)\n" % fileBlockNames
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         for block in fileBlockNames:
             valid_nodes = set(blocksInfo.get(block, [])) - node_filter
@@ -629,7 +668,7 @@ class DBS3Reader(object):
 
         if not self.blockExists(fileBlockName):
             msg = "DBSReader.getFileBlockWithParents(%s): No matching data"
-            raise DBSReaderError(msg % fileBlockName)
+            raise DBSReaderError(msg % fileBlockName) from None
 
         result = {"PhEDExNodeNames": [],  # FIXME: we better get rid of this line!
                   "Files": self.listFilesInBlockWithParents(fileBlockName)}
@@ -663,7 +702,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.blockToDatasetPath(%s)\n" % blockName
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         if blocks == []:
             return None
@@ -686,7 +725,7 @@ class DBS3Reader(object):
         except dbsClientException as ex:
             msg = "Error in DBSReader: dbsApi.listBlocks(dataset=%s)\n" % datasetName
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
         if not blocksInfo:  # no data location from dbs
             return list(locations)
@@ -703,17 +742,17 @@ class DBS3Reader(object):
         and datasets unknown to DBS. Otherwise None is returned.
         """
         if pathName in ("", None):
-            raise DBSReaderError("Invalid Dataset Path name: => %s <=" % pathName)
+            raise DBSReaderError("Invalid Dataset Path name: => %s <=" % pathName) from None
         else:
             try:
                 result = self.dbs.listDatasets(dataset=pathName, dataset_access_type='*')
                 if len(result) == 0:
-                    raise DBSReaderError("Dataset %s doesn't exist in DBS %s" % (pathName, self.dbsURL))
+                    raise DBSReaderError("Dataset %s doesn't exist in DBS %s" % (pathName, self.dbsURL)) from None
             except (dbsClientException, HTTPError) as ex:
                 msg = "Error in "
                 msg += "DBSReader.checkDatasetPath(%s)\n" % pathName
                 msg += "%s\n" % formatEx3(ex)
-                raise DBSReaderError(msg)
+                raise DBSReaderError(msg) from None
         return
 
     def checkBlockName(self, blockName):
@@ -721,7 +760,7 @@ class DBS3Reader(object):
          _checkBlockName_
         """
         if blockName in ("", "*", None):
-            raise DBSReaderError("Invalid Block name: => %s <=" % blockName)
+            raise DBSReaderError("Invalid Block name: => %s <=" % blockName) from None
 
     def getFileListByDataset(self, dataset, validFileOnly=1, detail=True):
 
@@ -740,7 +779,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.getFileListByDataset(%s)\n" % dataset
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
     def listDatasetParents(self, childDataset):
         """
@@ -753,7 +792,7 @@ class DBS3Reader(object):
             msg = "Error in "
             msg += "DBSReader.listDatasetParents(%s)\n" % childDataset
             msg += "%s\n" % formatEx3(ex)
-            raise DBSReaderError(msg)
+            raise DBSReaderError(msg) from None
 
     # def getListFilesByLumiAndDataset(self, dataset, files):
     #     "Unsing pycurl to get all the child parents pair for given dataset"
@@ -782,6 +821,9 @@ class DBS3Reader(object):
         :return: set of parent files for childLFN
         """
         fInfo = self.dbs.listFileLumiArray(logical_file_name=childLFNs)
+        if self.parallel:
+            return dbsParentFilesGivenParentDataset(self.dbsURL, parentDataset, fInfo)
+
         parentFiles = defaultdict(set)
         for f in fInfo:
             pFileList = self.dbs.listFiles(dataset=parentDataset, run_num=f['run_num'], lumi_list=f['lumi_section_num'])
@@ -911,7 +953,8 @@ class DBS3Reader(object):
                 numFiles = self.findAndInsertMissingParentage(blockName, parentFullInfo, insertFlag=insertFlag)
                 self.logger.debug("%s file parentage added for block %s", numFiles, blockName)
             except Exception as ex:
-                self.logger.exception("Parentage updated failed for block %s", blockName)
+                self.logger.exception(
+                        "Parentage updated failed for block %s with error %s", blockName, str(ex))
                 failedBlocks.append(blockName)
 
         return failedBlocks