Fixes #11098 

#### Status
ready

#### Description
Use concurrent features to speed up `listDatasetFileDetails` API. In my local setup I achieve speed up by factor of 5 or more for a specific dataset. Here are benchmark numbers for using `/VBF1Parked/Run2012D-v1/RAW` dataset. This dataset contains 594 blocks. The current codebase takes approximately 190 seconds to fetch its parents and file lumis. Using proposed solution I achieve the following numbers:
- 37 seconds using 50 concurrent tasks
- 41 seconds using 100 concurrent tasks
- 90 seconds using 10 concurrent tasks

Internally, I used `requests` python library instead of `pycurl_manager.py`. The latter is not suitable for concurrent execution since it is not thread safe as it holds global curl object. As such curl options are set on first tasks, but can't be changed (since code set them up) in others until first task is completed. The `requests` library does not depend on global curl object and curl library itself, and it is suitable for concurrent execution of multiple URL calls.

This PR only addresses speed up of single  `listDatasetFileDetails` DBS3Reader API but other APIs where multiple (sequential) calls to DBS are made can benefit from this approach.

#### Is it backward compatible (if not, which system it affects?)
YES

#### Related PRs
https://github.com/dmwm/dbs2go/issues/5

#### External dependencies / deployment changes
relies on Python `requests` library