@vkuznet 2022-04-15
@amaltaro I explained in description that `pycurl_manager.py` can't be used in concurrent futures since it locks curl object. The only way to use `pycurl_manager` is to use `multi_getdata` and it is on my todo list.