@vkuznet 2022-04-15
ok, I refactored the code using `pycurl_manager.py` `multi_getdata` and results are even more impressive. I achieve 10+ improvements for this API. Here is new structure:
- WMCore/Services/DBS/DBSUtils.py contains new common code
- WMCore/Services/DBS/DBS3Reader.py now can use either approach, tested with `requests` and `pycurl_manager.py` libraries
- I borrowed code from WMCore/MicroService/Tools/Common.py which seems to be common between these two code-bases, therefore I suggest to re-factor the code again to put this common code elsewhere. Please see appropriate comment in WMCore/Services/DBS/DBSUtils.py

Bottom line, I would like to keep around concurrent.features code since it demonstrates new paradigm on how concurrent programming should be done. This concept can be applied to different parts of WMCore which may benefit from it. The `requests` library used in original codebase is just an example and I suggest to keep it around, but it is not required now after my second implementation with `multi_getdata` from `pycurl_manager.py`