@vkuznet 2022-04-15
New best value I achieved with `multi_getdata` is 16 seconds, which is almost factor of 12 better than current implementation based on sequential programming pattern.