@vkuznet 2022-04-18
I re-factor the code and generalize it. The new API `dbsParallelApi` will allow to be used by different APIs in DBS3Reader (see full list in https://github.com/dmwm/WMCore/issues/11098). I added new attribute `parallel` into constructor of DBS3Reader such that class can execute either linear (sequential) or concurrent (parallel) DBS APis workflows. The code now can be further expanded to speed-up other APIs.