@vkuznet 2022-04-26
@amaltaro I added one _**integration**_ test to test functionality of `listDatasetFileDetails` DBS API through sequential and parallel execution. I found small dataset (in terms of blocks/files) in production DBS DB and used for the test. Please review and let me know if it is appropriate.