@vkuznet 2022-04-27
@amaltaro , I adjusted all comments you made except codes. The codes 200-300 have special meaning, e.g. 201 Created  is used to create resource, and we are not creating anything, 202 Accepted The request has been received but not yet acted upon, so when we query the data we should receive the data, etc. Please do not mix codes with actual behavior. Each code mans something and our action ONLY to receive the data which falls into code 200. All other codes are not appropriate for our action. That's why code should not use them as correct condition.