@vkuznet 2022-05-05
@amaltaro , I made necessary changes and replied to all your comments (some of them were unclear and I left my questions). Meanwhile, I run code through `autopep8`, and verified that Jenkins reports 10 score for new module. I also complement the code with integration tests. Please review again.